AIST3 RC1. Reactivated PvT1GateReaver. Added PvT2BaseReaver. Fixed most Shuttle waffling. Made PvT build more likely to get Reavers and upgrade shuttles. Introduced more PvT Reaver reactions

+++ src/Micro/Agency/Agency.scala
@@ -34,7 +34,7 @@ class Agency {
         .map(_.agent)
         .toVector
         .sortBy(_.unit.frameDiscovered) // Start with a stable order
-        .sortBy(_.unit.matchups.framesOfSafety) // Units in trouble get first dibs on things
+        .sortBy(_.unit.matchups.framesOfSafety.toInt) // Units in trouble get first dibs on things
         .sortBy(_.unit.unitClass.isTransport) // Make transports go after their passengers so they know what passengers want
     }
